Display player links for search query
