+++ src/hooks/useBasketballRef.js
@@ -1,0 +1,5 @@
+import { useFetch } from "./useFetch";
+
+export const useBasketballRef = (apiUrl, query, options) => {
+  return useFetch(`${apiUrl}/${query}`, options);
+}
